Move tests for generator-based coroutine support to their own module
